% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_advice.R
\name{catch_advice}
\alias{catch_advice}
\title{catch_advice}
\usage{
catch_advice(Cy, TBA, SAM, T = NULL)
}
\arguments{
\item{Cy}{The most recent annual catch, or average catch from multiple years.}

\item{TBA}{The Targeted Biomass Adjustment calculated using the \code{TBA} function.}

\item{SAM}{The Size Adherence Multiplier calculated using the \code{SAM} function.}

\item{T}{Threshold for limiting changes to catch advice. Default is \code{0.2} (i.e., +/- 20\%).}
}
\value{
Annual catch advice. Units are equal to those input for Cy.
}
\description{
The function calculates catch advice using the SlotLim framework.
}
\examples{
Cy <- 1000
TBA <- 1.1
SAM <- 0.9
T <- 0.2

Ay <- catch_advice(Cy, TBA, SAM, T)[[1]]

}
\seealso{
\code{\link{TBA}} # for the Targeted Biomass Adjustment

\code{\link{SAM}} # for the Size Adherence Multiplier
}
