% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSAM.R
\name{plotSAM}
\alias{plotSAM}
\title{plotSAM}
\usage{
plotSAM(L_2.5, L_97.5, MinLS, MaxLS, res = 1)
}
\arguments{
\item{L_2.5}{The 2.5th percentile of the catch from fishery-dependent length-frequency data.}

\item{L_97.5}{The 97.5th percentile of the catch from fishery-dependent length-frequency data.}

\item{MinLS}{The minimum landing size.}

\item{MaxLS}{The maximum landing size.}

\item{res}{Resolution for generating observations. Default is \code{1}. Smaller resolution values increase smoothness but may result in longer computation times.}
}
\value{
A \code{ggplot2} plot object
}
\description{
This function produces a plot visualizing the behavior of the Size Adherence Multiplier (SAM) equation, upon which the input data are plotted.
The plot can be saved as an object and further customized using \code{ggplot2} functions. Recommended input units are cm. For larger species, reduce the resolution to speed up plotting.
}
\examples{
L_2.5 <- 35      # 2.5th percentile
L_97.5 <- 50     # 97.5th percentile
MinLS <- 30      # Minimum landing size
MaxLS <- 50      # Maximum landing size
res <- 0.5       # Resolution

p <- plotSAM(L_2.5, L_97.5, MinLS, MaxLS, res)
p + ggplot2::theme(text = ggplot2::element_text(size = 10))

}
\seealso{
\code{\link{SAM}} # For the Size Adherence Multiplier
}
