% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTBA.R
\name{plotTBA}
\alias{plotTBA}
\title{plotTBA}
\usage{
plotTBA(P_targeted, rB)
}
\arguments{
\item{P_targeted}{The proportion of the size range targeted by harvest slot limits, calculated using the \code{prop_target} function.}

\item{rB}{The proportional rate of change in a biomass index, calculated using the \code{rB} function.}
}
\value{
A \code{ggplot2} plot object
}
\description{
This function produces a plot visualizing the behavior of the Targeted Biomass Adjustment (TBA) equation, upon which the input data are plotted.
The plot can be saved as an object and further customized using \code{ggplot2} functions.
}
\examples{
P_targeted <- 0.5
rB <- -0.5
p <- plotTBA(P_targeted, rB)
p + ggplot2::theme(text = ggplot2::element_text(size = 20))

}
\seealso{
\code{\link{prop_target}} # for the proportion targeted by HSL

\code{\link{rB}} # for the proportional rate of change in a biomass

\code{\link{TBA}} # for TBA function
}
