% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_target.R
\name{prop_target}
\alias{prop_target}
\title{prop_target}
\usage{
prop_target(MinLS, MaxLS, L_max)
}
\arguments{
\item{MinLS}{The minimum landing size.}

\item{MaxLS}{The maximum landing size.}

\item{L_max}{The maximum observed length.}
}
\value{
The proportion of the size range targeted by harvest slot limits.
}
\description{
The function calculates the proportion of the size range targeted by harvest slot limits.
}
\examples{
MinLS <- 50     # Minimum landing size
MaxLS <- 100    # Maximum landing size
L_max <- 200    # Maximum observed length

P_targeted <- prop_target(MinLS, MaxLS, L_max)
print(P_targeted)

}
