% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBA.R
\name{TBA}
\alias{TBA}
\title{TBA}
\usage{
TBA(P_targeted, rB)
}
\arguments{
\item{P_targeted}{The proportion of abundance targeted by harvest slot limits, calculated using the \code{prop_target} function.}

\item{rB}{The proportional rate of change in a biomass index, calculated using the \code{rB} function.}
}
\value{
The targeted biomass adjustment. When TBA > 1, catch advice increases, When TBA < 1, catch advice decreases.
}
\description{
This function calculates the Targeted Biomass Adjustment (TBA), which dampens the influence of biomass indices on catch advice when the proportion of abundance targeted by harvest slot limits is small.
}
\examples{
P_targeted <- 0.5
rB <- -0.5

TBA(P_targeted, rB)

}
\seealso{
\code{\link{prop_target}} # for the proportion of abundance targeted by HSL

\code{\link{rB}} # for the proportional rate of change in a biomass
}
