% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_target.R
\name{prop_target}
\alias{prop_target}
\title{prop_target}
\usage{
prop_target(MinLS, MaxLS, M = NULL, L_inf, K, t_0, t_max, L_min = 1e-05)
}
\arguments{
\item{MinLS}{The minimum landing size.}

\item{MaxLS}{The maximum landing size.}

\item{M}{Instantaneous natural mortality rate.}

\item{L_inf}{Asymptotic length.}

\item{K}{Juvenile growth coefficient.}

\item{t_0}{Hypothetical age at which length is zero.}

\item{t_max}{Maximum age.}

\item{L_min}{To account for non-zero length-at-age-0, survivorship is modeled from 1*10^-5.}
}
\value{
The proportion of abundance targeted by harvest slot limits.
}
\description{
The function estimates the proportion of abundance targeted by harvest slot limits from a length-structured survivorship curve. If left blank, M is calculated using t_max as per Hoenig's equation modified by Then et al. (2015): Then, A. Y., Hoenig, J. M., Hall, N. G., Hewitt, D. A. & Handling editor: Ernesto Jardim. (2015). Evaluating the predictive performance of empirical estimators of natural mortality rate using information on over 200 fish species. ICES Journal of Marine Science 72, 82 – 92.
}
\examples{
MinLS <- 120
MaxLS <- 240
M <- 0.19
L_inf <- 405
K <- 0.118
t_0 <- -1.83
t_max <- 34
prop_target(MinLS, MaxLS, M, L_inf, K, t_0, t_max)

}
